requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
sentence-transformers==2.7.0
streamlit==1.31.0
python-dotenv==1.0.1
openai>=1.0.0
pandas==2.2.1
google-search-results